Implemented authorization in all microservices. Updated the gateway.

+++ Clothy.Services/Clothy.ReviewService/Clothy.ReviewService.Application/Features/Questions/Commands/DeleteAnswer/DeleteAnswerToQuestionCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 using Clothy.ReviewService.Application.Interfaces.Commands;
@@ -24,6 +25,11 @@ namespace Clothy.ReviewService.Application.Features.Questions.Commands.DeleteAns
         {
             Question? question = await questionRepository.GetByIdAsync(request.QuestionId, cancellationToken);
             if (question == null) throw new NotFoundException($"Question with ID {request.QuestionId} not found!");
+
+            Answer? answer = question.Answers.FirstOrDefault(tempAnswer => tempAnswer.Id == request.AnswerId);
+            if (answer == null) throw new NotFoundException($"Answer with ID {request.AnswerId} not found!");
+
+            if (question.User.UserId != request.UserId && !request.IsAdmin) throw new ForbiddenException("You are not allowed to delete this answer");
 
             await questionRepository.DeleteAnswerAsync(request.QuestionId, request.AnswerId, cancellationToken);
             return Unit.Value;
